Got java working as much as c++ except for an issue with enum casting

+++ cpp/include/ctre/phoenix/CustomParamConfiguration.h
@@ -4,8 +4,8 @@
 namespace ctre {
 namespace phoenix {
 struct CustomParamConfiguration {
-	double CustomParam_0;	
-	double CustomParam_1;	
+	int CustomParam_0;	
+	int CustomParam_1;	
 	CustomParamConfiguration() :
 		CustomParam_0(0),
 		CustomParam_1(0)
